refactor(table): :recycle: move controls and handlers into hook
