+++ app/configurator/table/table-chart-configurator.tsx
@@ -1,19 +1,10 @@
 import { Trans } from "@lingui/macro";
 import { Theme, Typography } from "@mui/material";
 import { makeStyles } from "@mui/styles";
-import { useCallback, useMemo, useState } from "react";
-import {
-  DragDropContext,
-  DraggableLocation,
-  OnDragStartResponder,
-} from "react-beautiful-dnd";
+import { DragDropContext } from "react-beautiful-dnd";
 
 import { Loading } from "@/components/hint";
-import { TableFields } from "@/config-types";
-import {
-  ConfiguratorStateConfiguringChart,
-  getChartConfig,
-} from "@/configurator";
+import { ConfiguratorStateConfiguringChart } from "@/configurator";
 import { TabDropZone } from "@/configurator/components/chart-controls/drag-and-drop-tab";
 import {
   ControlSection,
@@ -22,17 +13,10 @@ import {
 } from "@/configurator/components/chart-controls/section";
 import { ChartTypeSelector } from "@/configurator/components/chart-type-selector";
 import { AnnotatorTabField } from "@/configurator/components/field";
-import { useOrderedTableColumns } from "@/configurator/components/ui-helpers";
-import {
-  isConfiguring,
-  useConfiguratorState,
-} from "@/configurator/configurator-state";
-import { moveFields } from "@/configurator/table/table-config-state";
-import {
-  useComponentsWithHierarchiesQuery,
-  useDataCubeMetadataQuery,
-} from "@/graphql/query-hooks";
-import { useLocale } from "@/locales/use-locale";
+
+import { useOrderedTableColumns } from "../components/ui-helpers";
+
+import { useTableChartController } from "./table-chart-configurator.hook";
 
 const useStyles = makeStyles((theme: Theme) => ({
   emptyGroups: {
@@ -62,159 +46,97 @@ export const ChartConfiguratorTable = ({
 }: {
   state: ConfiguratorStateConfiguringChart;
 }) => {
-  const locale = useLocale();
-  const [, dispatch] = useConfiguratorState(isConfiguring);
-  const chartConfig = getChartConfig(state);
-  const variables = {
-    iri: chartConfig.dataSet,
-    sourceType: state.dataSource.type,
-    sourceUrl: state.dataSource.url,
-    locale,
-  };
-  const [{ data: metadata }] = useDataCubeMetadataQuery({ variables });
-  const [{ data: components }] = useComponentsWithHierarchiesQuery({
-    variables,
-  });
+  const {
+    metaData,
+    currentDraggableId,
+    chartConfig,
+    onDragEnd,
+    onDragStart,
+    handleMove,
+  } = useTableChartController(state);
 
-  const metaData = useMemo(() => {
-    return metadata?.dataCubeByIri && components?.dataCubeByIri
-      ? {
-          ...metadata.dataCubeByIri,
-          ...components.dataCubeByIri,
-        }
-      : null;
-  }, [metadata?.dataCubeByIri, components?.dataCubeByIri]);
+  const fieldsArray = useOrderedTableColumns(chartConfig.fields);
 
-  const [currentDraggableId, setCurrentDraggableId] = useState<string | null>(
-    null
-  );
-
-  const onDragEnd = useCallback(
-    ({
-      source,
-      destination,
-    }: {
-      source: DraggableLocation;
-      destination?: DraggableLocation | null;
-    }) => {
-      setCurrentDraggableId(null);
-
-      if (!destination || chartConfig.chartType !== "table" || !metaData) {
-        return;
-      }
-
-      const newChartConfig = moveFields(chartConfig, {
-        source,
-        destination,
-      });
-
-      dispatch({
-        type: "CHART_CONFIG_REPLACED",
-        value: {
-          chartConfig: newChartConfig,
-          dataSetMetadata: metaData,
-        },
-      });
-    },
-    [chartConfig, dispatch, metaData]
-  );
-
-  const onDragStart = useCallback<OnDragStartResponder>(({ draggableId }) => {
-    setCurrentDraggableId(draggableId);
-  }, []);
-
-  const fields = chartConfig.fields as TableFields;
-  const fieldsArray = useOrderedTableColumns(fields);
-
-  if (metaData) {
-    const groupFields = [...fieldsArray.filter((f) => f.isGroup)];
-    const columnFields = [...fieldsArray.filter((f) => !f.isGroup)];
-
-    const currentDraggedField =
-      currentDraggableId !== null ? fields[currentDraggableId] : null;
-    const isGroupsDropDisabled =
-      currentDraggedField?.componentType === "NumericalMeasure";
-
-    return (
-      <>
-        <ControlSection collapse>
-          <SubsectionTitle titleId="controls-design" gutterBottom={false}>
-            <Trans id="controls.select.chart.type">Chart Type</Trans>
-          </SubsectionTitle>
-          <ControlSectionContent px="small" gap="none">
-            <ChartTypeSelector
-              showHelp={false}
-              state={state}
-              chartKey={chartConfig.key}
-              sx={{ mt: 2 }}
-            />
-          </ControlSectionContent>
-        </ControlSection>
-        <ControlSection collapse>
-          <SubsectionTitle gutterBottom={false}>
-            <Trans id="controls.section.tableoptions">Table Options</Trans>
-          </SubsectionTitle>
-          <ControlSectionContent
-            px="small"
-            gap="none"
-            role="tablist"
-            aria-labelledby="controls-settings"
-          >
-            <AnnotatorTabField
-              key={"settings"}
-              value={"table-settings"}
-              icon="settings"
-              mainLabel={<Trans id="controls.table.settings">Settings</Trans>}
-            />
-            <AnnotatorTabField
-              key={"sorting"}
-              value={"table-sorting"}
-              icon="sort"
-              mainLabel={<Trans id="controls.table.sorting">Sorting</Trans>}
-            />
-          </ControlSectionContent>
-        </ControlSection>
-        <DragDropContext
-          onDragEnd={(result) =>
-            onDragEnd({
-              source: result.source,
-              destination: result.destination,
-            })
-          }
-          onDragStart={onDragStart}
-        >
-          <TabDropZone
-            id="groups"
-            title={<Trans id="controls.section.groups">Groups</Trans>}
-            metaData={metaData}
-            items={groupFields}
-            isDropDisabled={isGroupsDropDisabled}
-            emptyComponent={<EmptyGroups />}
-            onUp={(_index) => {}} // TODO: implement
-            onDown={(_index) => {}} // TODO: implement
-          />
-
-          <TabDropZone
-            id="columns"
-            title={<Trans id="controls.section.columns">Columns</Trans>}
-            metaData={metaData}
-            items={columnFields}
-            onUp={(newIndex) => {
-              if (newIndex < 0) {
-                return;
-              }
-
-              onDragEnd({
-                source: { droppableId: "columns", index: newIndex + 1 },
-                destination: { droppableId: "columns", index: newIndex },
-              });
-            }} // TODO: implement
-            onDown={(_index) => {}} // TODO: implement
-          />
-        </DragDropContext>
-      </>
-    );
-  } else {
+  if (!metaData) {
     return <Loading />;
   }
+
+  const groupFields = [...fieldsArray.filter((f) => f.isGroup)];
+  const columnFields = [...fieldsArray.filter((f) => !f.isGroup)];
+
+  const currentDraggedField =
+    currentDraggableId !== null ? chartConfig.fields[currentDraggableId] : null;
+  const isGroupsDropDisabled =
+    currentDraggedField?.componentType === "NumericalMeasure";
+
+  return (
+    <>
+      <ControlSection collapse>
+        <SubsectionTitle titleId="controls-design" gutterBottom={false}>
+          <Trans id="controls.select.chart.type">Chart Type</Trans>
+        </SubsectionTitle>
+        <ControlSectionContent px="small" gap="none">
+          <ChartTypeSelector
+            showHelp={false}
+            state={state}
+            chartKey={chartConfig.key}
+            sx={{ mt: 2 }}
+          />
+        </ControlSectionContent>
+      </ControlSection>
+      <ControlSection collapse>
+        <SubsectionTitle gutterBottom={false}>
+          <Trans id="controls.section.tableoptions">Table Options</Trans>
+        </SubsectionTitle>
+        <ControlSectionContent
+          px="small"
+          gap="none"
+          role="tablist"
+          aria-labelledby="controls-settings"
+        >
+          <AnnotatorTabField
+            key={"settings"}
+            value={"table-settings"}
+            icon="settings"
+            mainLabel={<Trans id="controls.table.settings">Settings</Trans>}
+          />
+          <AnnotatorTabField
+            key={"sorting"}
+            value={"table-sorting"}
+            icon="sort"
+            mainLabel={<Trans id="controls.table.sorting">Sorting</Trans>}
+          />
+        </ControlSectionContent>
+      </ControlSection>
+      <DragDropContext
+        onDragEnd={(result) =>
+          onDragEnd({
+            source: result.source,
+            destination: result.destination,
+          })
+        }
+        onDragStart={onDragStart}
+      >
+        <TabDropZone
+          id="groups"
+          title={<Trans id="controls.section.groups">Groups</Trans>}
+          metaData={metaData}
+          items={groupFields}
+          isDropDisabled={isGroupsDropDisabled}
+          emptyComponent={<EmptyGroups />}
+          onUp={handleMove(-1, "groups")}
+          onDown={handleMove(1, "groups")}
+        />
+
+        <TabDropZone
+          id="columns"
+          title={<Trans id="controls.section.columns">Columns</Trans>}
+          metaData={metaData}
+          items={columnFields}
+          onUp={handleMove(-1, "columns")}
+          onDown={handleMove(1, "columns")}
+        />
+      </DragDropContext>
+    </>
+  );
 };
